+++ Assets/Scripts/Tuio/Tuio11/Tuio11ObjectBehaviour.cs
@@ -3,6 +3,7 @@ using TuioNet.Server;
 using TuioNet.Tuio11;
 using TuioSimulator.Input;
 using TuioSimulator.Tuio.Common;
+using TuioSimulator.Utils;
 using UnityEngine;
 using UnityEngine.EventSystems;
 using Utils;
